+++ CHANGELOG.md
@@ -11,10 +11,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Breaking changes
 
 * The type of `price` field in `Posting` has changed to `Option<PostingPrice<'a, D>>`.
+* The type `MetadataValue` no longer implement `Eq`
 
 ### Added
 
 * Support for total price in posting (`@@` syntax)
+* implement `Clone` for all types
+* implement `Copy`, `Eq`, `Ord` and `Hash` on `Account` and `Currency`
 
 
 ## [1.0.0-alpha.1] - 2023-05-21